2020-1/5°ADS/(Equipe 2) segurancadainformacao/requirements.txt
alembic==1.4.2
click==7.1.1
Flask==1.1.1
Flask-Login==0.5.0
Flask-Migrate==2.5.2
Flask-Script==2.0.6
Flask-SQLAlchemy==2.4.1
itsdangerous==1.1.0
Jinja2==2.11.1
Mako==1.1.2
MarkupSafe==1.1.1
pycryptodome==3.9.7
pycryptodomex==3.9.7
python-dateutil==2.8.1
python-editor==1.0.4
six==1.14.0
SQLAlchemy==1.3.15
Werkzeug==1.0.0
